requirements.txt
aiogram==3.13.0
pillow==10.4.0
python-dotenv==1.0.1